refactor(api): migrate route PATCH /api/admin/users/{id}
